Add adaptive nature to the main navigation.
This uses progressive enhancement, meaning the base style has the nav small and
media queries bring the size larger. The math used is no longer relevant and it
should be changed, but that isn't the priority at the moment. Note that in a
browser without media queries (ahem, IE) this will not work out very well. There
is a polyfill to help. In general, browser support will be considered in later
commits.