Add vote counter wheels and social links.
This is absolutely not cross-browser compatible, so it needs to be refactored.
Since the entire schema relies on JavaScript for rotation, it may be worth
hiding the wheels by default and using JavaScript to show them conditional on
the browser's ability to transform. Resolves #10.

Also, add social links. They don't have a caption, and only show up if the
submission has the relevant information available. Resolves #18.